+++ tsai/callback/all.py
@@ -1,5 +1,5 @@
 from .core import *
 from .noisy_student import *
 from .gblend import *
-from .TSBERT import *
+from .MVP import *
 from .PredictionDynamics import *